requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5

Flask-WTF==1.2.2
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.1

Werkzeug==2.3.7
python-dotenv==1.0.0
qrcode==7.4.2
Pillow>=10.4,<12

# CORREÇÃO: Atualizar WeasyPrint para corrigir erro de compatibilidade com pydyf
WeasyPrint>=62.0